Adding benchmarks for embedded/non-embedded json within plaintext
